+++ Tests/SentryTests/Performance/SentryTracerTests.swift
@@ -235,6 +235,28 @@ class SentryTracerTests: XCTestCase {
         assertAppStartMeasurementNotPutOnTransaction()
     }
     
+    func testFinish_WithUnfinishedChildren() {
+        CurrentDate.setCurrentDateProvider(DefaultCurrentDateProvider.sharedInstance())
+        let sut = fixture.getSut(waitForChildren: false)
+        let child1 = sut.startChild(operation: fixture.transactionOperation)
+        let child2 = sut.startChild(operation: fixture.transactionOperation)
+        let child3 = sut.startChild(operation: fixture.transactionOperation)
+        child2.finish()
+        sut.finish()
+        
+        XCTAssertTrue(child1.isFinished)
+        XCTAssertEqual(child1.context.status, .deadlineExceeded)
+        XCTAssertEqual(sut.timestamp, child1.timestamp)
+        
+        XCTAssertTrue(child2.isFinished)
+        XCTAssertEqual(child2.context.status, .ok)
+        XCTAssertNotEqual(sut.timestamp, child2.timestamp)
+        
+        XCTAssertTrue(child3.isFinished)
+        XCTAssertEqual(child3.context.status, .deadlineExceeded)
+        XCTAssertEqual(sut.timestamp, child3.timestamp)
+    }
+    
     // Although we only run this test above the below specified versions, we expect the
     // implementation to be thread safe
     @available(tvOS 10.0, *)