+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat: Include unfinished spans in transactions (#1592)
 - build: Disable NSAssertions for Release Builds (#1545)
 
 ## 7.7.0